requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
fastapi==0.61.2
h11==0.11.0
paramiko==2.7.2
pycparser==2.20
pydantic==1.7.2
PyMySQL==0.10.1
PyNaCl==1.4.0
six==1.15.0
starlette==0.13.6
uvicorn==0.12.3
websockets==8.1
sqlalchemy==1.3.20
pyjwt==1.7.1
passlib==1.7.4
python-multipart==0.0.5
aiofiles==0.6.0
pycryptodome==3.9.9
pytz==2020.4
sqlparse==0.4.1
cassandra-driver==3.25.0
cqlsh==6.0.0